1.16.5 Merge of 2.1.0 Update Candidate

+++ src/main/java/zmaster587/advancedRocketry/entity/fx/InverseTrailFluid.java
@@ -1,6 +1,5 @@
 package zmaster587.advancedRocketry.entity.fx;
 
-
 import net.minecraft.client.particle.IAnimatedSprite;
 import net.minecraft.client.particle.IParticleFactory;
 import net.minecraft.client.particle.Particle;
@@ -26,10 +25,10 @@ public class InverseTrailFluid extends InverseTrailFx {
 		
 		float intensity = ((float)(Math.random())*0.3f) + 0.7f;
 		
-		this.particleRed = (float) (((color >> 16) & 0xFF)/255f) *intensity;
-		this.particleGreen = (float) (((color >> 8) & 0xFF)/255f)*intensity;
-		this.particleBlue = (float) (((color& 0xFF)/255f))*intensity;
-		
+		this.particleRed = (((color >> 16) & 0xFF)/255f) *intensity;
+		this.particleGreen = (((color >> 8) & 0xFF)/255f) *intensity;
+		this.particleBlue = ((color& 0xFF)/255f) *intensity;
+
         this.maxAge = time;
         this.particleScale = (float) (this.rand.nextFloat() * 0.25F) + 0.75f;
 	}
