+++ src/main/java/zmaster587/advancedRocketry/entity/fx/TrailFx.java
@@ -26,7 +26,7 @@ public class TrailFx extends InverseTrailFx {
         this.particleGreen = .4F + chroma;
         this.particleBlue = .4F + chroma;
         this.setSize(0.12F, 0.12F);
-        this.particleScale = (float)(this.rand.nextFloat() * 0.6F + 2F);
+        this.particleScale = this.rand.nextFloat() * 0.6F + 2F;
         this.motionX = motx;
         this.motionY = moty;
         this.motionZ = motz;
